perf: optimise `Intents` usage. (#17)
This commit uses the intent integer flags appropriate for all operations ran in Astro. `Intents.DEFAULT` is an inefficient way of allowing the library to continue dispatching events which can serve as a computational burden. For most use cases, it's fine to pass this default but in our situation we want to make Astro as fast as humanly possible.